refactor(agent): decompose profile record production
