+++ packages/agent/src/system-records/agent-profile-producer-signing-v1.ts
@@ -1,0 +1,65 @@
+import {
+  buildSystemRecordSignatureMessageV1,
+  canonicalizeSignedSystemRecordEnvelopeV1,
+  verifySignedSystemRecordEnvelopeV1,
+  type SignedAgentProfileHeadEnvelopeV1,
+} from '@origintrail-official/dkg-core/system-record-v1';
+
+import type { CreateAgentProfileProducerOptionsV1 } from './agent-profile-producer-contract-v1.js';
+import type { AgentProfileProductionPreparationV1 } from './agent-profile-producer-preparation-v1.js';
+
+export interface SignedAgentProfileProductionV1 {
+  readonly envelope: SignedAgentProfileHeadEnvelopeV1;
+  readonly envelopeBytes: Uint8Array;
+}
+
+export async function signAgentProfileProductionV1(
+  options: CreateAgentProfileProducerOptionsV1,
+  preparation: AgentProfileProductionPreparationV1,
+  signal: AbortSignal,
+): Promise<SignedAgentProfileProductionV1> {
+  const [peerSignature, evmSignature] = await Promise.all([
+    options.peerSigner.sign(
+      buildSystemRecordSignatureMessageV1(
+        preparation.head,
+        preparation.headDigest,
+        'peer',
+      ),
+    ),
+    options.evmSigner.signMessage(
+      buildSystemRecordSignatureMessageV1(
+        preparation.head,
+        preparation.headDigest,
+        'current-evm',
+      ),
+    ),
+  ]);
+  signal.throwIfAborted();
+  const envelope: SignedAgentProfileHeadEnvelopeV1 = {
+    object: preparation.head,
+    objectDigest: preparation.headDigest,
+    signatures: Object.freeze([
+      Object.freeze({
+        role: 'peer',
+        suite: 'ed25519-v1',
+        signer: options.peerSigner.peerId,
+        evidence: Object.freeze({ kind: 'none' }),
+        signature: Buffer.from(peerSignature).toString('base64url'),
+      }),
+      Object.freeze({
+        role: 'current-evm',
+        suite: 'eip191-personal-sign-digest-v1',
+        signer: options.evmSigner.address,
+        evidence: Object.freeze({ kind: 'none' }),
+        signature: evmSignature,
+      }),
+    ]),
+  };
+  if (!await verifySignedSystemRecordEnvelopeV1(envelope)) {
+    throw new Error('new profile head signature verification failed');
+  }
+  return Object.freeze({
+    envelope,
+    envelopeBytes: canonicalizeSignedSystemRecordEnvelopeV1(envelope),
+  });
+}
